Made whale get damage from the sun cuz whales dont die of not having air
Whales irl die if they're exposed to the sun for too long as it burns
their flesh, not because of not having air. There are rare cases where
whales have died from not being in water too long and not being exposed
to the sun, but we dont talk about that.