FINAL COMMIT BEFORE UPDATE:
-Fixed bird imitating parrot sounds
-Permanently fixed reach and knockback
-Fixed tool stat config values not updating in-game
-Fixed a rendering bug
-Fixed critter cage bugging due to entities having custom names (being
name-tagged)
-Fixed AI completely breaking
-Fixed wasp server crash
-Fixed Ent not being able to pathfind or make one block jumps
-A bunch of useless classes that'll be removed once I have more time to
sort/refactor code
-Fixed a client config crash
-Added a client config
-Holy shit, the code really needs to be reformatted
-Fixed some random server crash that happened with the ender dragon
-Fixed ultimate pickaxe instantly disappearing when dropped
-Fixed ultimate pickaxe duplicating every block
-Fixed all other damage values of normal weapons/sets
-Fixed some item tier damage and efficiency values
-Fixed a world gen bug where sometimes the wasp nest would spawn very
deep underground
-Optimized fish AI
-Fixed whale death causing a render crash
-Fixed broken frog anims
-Fixed like 3 lava eel crashes
-Fixed lava eel AI to some degree
-Fixed crystal dimension AI somewhat (it's still slow)
-Fixed bird being brokey af
-Fixed dying with a big weapon causing rendering bugs when holding other
items
-Fixed a looting crash with ultimate sword
-Fixed robo lasers lagging the game to an unplayable state

FINAL FIXES NEEDED BEFORE PUBLIC RELEASE (not all are necessary):
-Ultimate fishing rod --Unnecessary
-Damage and reach values (only for big weapons, everything else is
fine)--Necessary
-Some AI tbh --Unnecessary
-Gazelle running animation broken --Necessary
-May need to have a final look at bird's behaviour

--NOTE: Yes, for anyone else not a contributor/developer looking at this
commit message, it is reasonable to think that it is unacceptable for
such an update to take so long, yet not provide some necessary fixes, AI
being the poster child for that. Your concerns are absolutely validated,
however in order to save time and not pronounce the mod to be dead, this
update focused a whole lot more on adding as much quality content as
possible. A (hopefully enforced) schedule will be established to
eliminate these annoying bugs soon enough by the next update, while not
taking several lifetimes to complete.

Thanks for being a supporter of the mod.